worker/requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
requests==2.32.3
msal==1.28.1
croniter==2.0.5
PyJWT==2.8.0
